Refactor deployments (#540)
