+++ src/index.ts
@@ -26,14 +26,13 @@ import * as commander from 'commander';
 import {
   apis,
   create,
-  deploy,
-  deployments,
   login,
   run,
-  undeploy,
 } from './commands';
 import clone from './commands/clone';
 import defaultCmd from './commands/defaultCmd';
+import deploy from './commands/deploy';
+import deployments from './commands/deployments';
 import help from './commands/help';
 import list from './commands/list';
 import logout from './commands/logout';
@@ -43,11 +42,10 @@ import pull from './commands/pull';
 import push from './commands/push';
 import setting from './commands/setting';
 import status from './commands/status';
+import undeploy from './commands/undeploy';
 import version from './commands/version';
 import versions from './commands/versions';
 import { PROJECT_NAME, handleError } from './utils';
-
-// const commander = require('commander');
 
 // CLI
